+++ checkers/testdata/timeCmpSimplify/positive_tests.go
@@ -1,4 +1,4 @@
-package timeComparisonSimplify
+package timeCmpSimplify
 
 import (
     "time"
